detailview
adw

+++ MyApplication2/app/src/main/java/com/example/simon/photojsl/DetailView.java
@@ -4,6 +4,7 @@ import android.content.ContentResolver;
 import android.content.ContentValues;
 import android.content.Context;
 import android.content.Intent;
+import android.graphics.Bitmap;
 import android.net.Uri;
 import android.os.Bundle;
 import android.provider.MediaStore;
@@ -15,7 +16,9 @@ import android.widget.ImageView;
 
 import java.io.File;
 import java.io.FileNotFoundException;
+import java.io.FileOutputStream;
 import java.io.IOException;
+import java.net.URI;
 
 public class DetailView extends AppCompatActivity {
     private File m_pic;
@@ -48,24 +51,27 @@ public class DetailView extends AppCompatActivity {
         //noinspection SimplifiableIfStatement
         switch(id){
             case R.id.action_delete:
-                m_pic.delete();
+                if(MainMenu.getAdapter().deleteListEntry(m_pic.getName())){
+                    m_pic.delete();
+                }
                 break;
             case R.id.action_send:
                 final Intent emailIntent = new Intent(android.content.Intent.ACTION_SEND);
-                emailIntent.setType("plain/text");
+//                emailIntent.setType("plain/text");
                 emailIntent.putExtra(android.content.Intent.EXTRA_SUBJECT, m_pic.getName());
                 emailIntent.putExtra(android.content.Intent.EXTRA_TEXT, "powered by JLS Software");
-                emailIntent.setType("image/png");
+
                 emailIntent.setFlags(Intent.FLAG_ACTIVITY_NEW_TASK);
-                emailIntent.putExtra(Intent.EXTRA_STREAM, Uri.fromFile(m_pic));
+                emailIntent.putExtra(Intent.EXTRA_STREAM, m_pic.getAbsolutePath());
+                emailIntent.setType("image/png");
                 startActivity(Intent.createChooser(emailIntent, "Send mail"));
                 break;
             case R.id.action_save:
-                try {
-                    Images.Media.insertImage(getContentResolver(), m_pic.getAbsolutePath(),m_pic.getName(), "JLS Picture");
-                } catch (FileNotFoundException e) {
-                    e.printStackTrace();
-                }
+                ContentValues values = new ContentValues();
+                values.put(Images.Media.DATE_TAKEN, System.currentTimeMillis());
+                values.put(Images.Media.MIME_TYPE, "image/png");
+                values.put(MediaStore.MediaColumns.DATA, m_pic.getAbsolutePath());
+                getApplicationContext().getContentResolver().insert(Images.Media.EXTERNAL_CONTENT_URI, values);
                 break;
         }
         return super.onOptionsItemSelected(item);
